requirements.txt
boto3
yt_dlp
pytube
psycopg2-binary
python-dotenv
requests